rename: dto 폴더 정리
