+++ src/main/java/gsc/backend/service/QuizService.java
@@ -6,9 +6,9 @@ import gsc.backend.domain.QuizAnswer;
 import gsc.backend.domain.User;
 import gsc.backend.domain.mapping.UserQuiz;
 import gsc.backend.dto.request.QuizAnswerRequestDTO;
-import gsc.backend.dto.response.QuizDataDTO;
-import gsc.backend.dto.response.QuizOptionsDTO;
-import gsc.backend.dto.response.QuizResponseDTO;
+import gsc.backend.dto.response.quiz.QuizDataDTO;
+import gsc.backend.dto.response.quiz.QuizOptionsDTO;
+import gsc.backend.dto.response.quiz.QuizResponseDTO;
 import gsc.backend.repository.*;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
